% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dr}
\alias{example_dr}
\title{Dose-response of cell lines to investigative compound treatment}
\format{
A tibble containing 240 rows and 4 variables:
\describe{
  \item{Cell}{Name of the cell line being treated}
  \item{Compound}{Name of the compound used}
  \item{Concentration (nM)}{The nanomolar (nM) concentration of the compound}
  \item{Viability}{The fractional effect of treatment on the measured viability of cells}
  ...
}
}
\source{
Author's dataset.
}
\usage{
example_dr
}
\description{
A dataset containing the viability response of two cell lines to 
different concentrations of series of investigative compounds. There
are two technical replicates for each compound/concentration combination.
}
\keyword{datasets}
