% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_HairEye}
\alias{example_HairEye}
\title{Distribution of hair and eye color and sex in 592 statistics students.}
\format{
A tibble containing 592 rows and 3 variables:
\describe{
  \item{Hair}{Hair color}
  \item{Eye}{Eye color}
  \item{Sex}{Sex}
  ...
}

A tibble containing 578 rows and 4 variables:
\describe{
  \item{weight}{a numeric vector giving the body weight of the chick (gm).}
  \item{Time}{a numeric vector giving the number of days since birth when the measurement was made.}
  \item{Chick}{an ordered factor with levels 18 < ... < 48 giving a unique identifier for the chick. The ordering of the levels groups chicks on the same diet together and orders them according to their final weight (lightest to heaviest) within diet.}
  \item{Diet}{a factor with levels 1, ..., 4 indicating which experimental diet the chick received.}
  ...
}
}
\source{
\url{http://www.datavis.ca/sas/vcd/catdata/haireye.sas}

Crowder, M. and Hand, D. (1990), Analysis of Repeated Measures, Chapman and Hall (example 5.3)
Hand, D. and Crowder, M. (1996), Practical Longitudinal Data Analysis, Chapman and Hall (table A.2)
Pinheiro, J. C. and Bates, D. M. (2000) Mixed-effects Models in S and S-PLUS, Springer.
}
\usage{
example_HairEye

example_HairEye
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
